TODO: add StridedSlice op
Fix bugs in: max/min/Equal/Range/Invert_permutation/StridedSlice/Tile operators

+++ src/nnfusion/core/kernels/cuda_gpu/kernels/range.hpp
@@ -20,7 +20,7 @@ namespace nnfusion
 
             private:
                 ngraph::Shape output_shape;
-                float start, limit, delta;
+                int start, limit, delta;
                 int range_num;
             };
         } // namespace cuda
